Hwago, LumosKomik (ID): Address review comments
