+++ lib-multisrc/astromangareader/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    alias(kei.plugins.multisrc)
+}
+
+keiyoushi {
+    baseVersionCode = 1
+    libVersion = "1.6"
+
+    deeplink {
+        path("/comic/..*")
+        path("/komik/..*")
+    }
+}
